docs: use react template for custom-radio-button
