+++ components/custom-checkbox/css/stories/default.stories.mdx
@@ -19,10 +19,7 @@ import "../index.scss";
 # Checkbox Component
 
 <Canvas>
-  <Story
-    name="Default"
-    args={defaultArgs}
-  >
+  <Story name="Default" args={defaultArgs}>
     {CustomCheckbox.bind({})}
   </Story>
 </Canvas>
