+++ components/custom-checkbox/web-component/stories/default.stories.mdx
@@ -18,10 +18,7 @@ import "../index.scss";
 # Checkbox Component
 
 <Canvas>
-  <Story
-    name="Default"
-    args={defaultArgs}
-  >
+  <Story name="Default" args={defaultArgs}>
     {CustomCheckbox.bind({})}
   </Story>
 </Canvas>
