+++ components/custom-radio-button/css/stories/default.stories.mdx
@@ -1,13 +1,13 @@
 <!-- @license CC0-1.0 -->
 
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { CustomRadioButton, defaultArgs } from "../template";
+import { argTypes, CustomRadioButton, defaultArgs } from "../story-template";
 import "../index.scss";
 
 <Meta
   id="css-custom-radio-button"
   title="CSS Component/Custom Radio Button"
-  argTypes={{}}
+  argTypes={argTypes}
   parameters={{
     status: {
       type: "WORK IN PROGRESS",
@@ -18,12 +18,7 @@ import "../index.scss";
 # Custom Radio Button Component
 
 <Canvas>
-  <Story
-    name="Default"
-    args={{
-      ...defaultArgs,
-    }}
-  >
+  <Story name="Default" args={defaultArgs}>
     {CustomRadioButton.bind({})}
   </Story>
 </Canvas>
